# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/lessons": {
      "get": {
        "operationId": "LessonsController_getAllLessons",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des leçons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtenir toutes les leçons publiées",
        "tags": [
          "Lessons"
        ]
      },
      "post": {
        "operationId": "LessonsController_createLesson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leçon créée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer une nouvelle leçon (instructeur)",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_getLessonById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de la leçon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtenir une leçon par ID",
        "tags": [
          "Lessons"
        ]
      },
      "put": {
        "operationId": "LessonsController_updateLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leçon mise à jour"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour une leçon",
        "tags": [
          "Lessons"
        ]
      },
      "delete": {
        "operationId": "LessonsController_deleteLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leçon supprimée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/my/created": {
      "get": {
        "operationId": "LessonsController_getMyLessons",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des leçons créées"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir mes leçons créées",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/{id}/publish": {
      "post": {
        "operationId": "LessonsController_publishLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leçon publiée"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Publier une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/{lessonId}/modules": {
      "post": {
        "operationId": "LessonsController_createModule",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Module créé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ajouter un module à une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/modules/{moduleId}/contents": {
      "post": {
        "operationId": "LessonsController_createContent",
        "parameters": [
          {
            "name": "moduleId",
            "required": true,
            "in": "path",
            "description": "ID du module",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contenu créé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Ajouter du contenu à un module",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/contents/{contentId}": {
      "get": {
        "operationId": "LessonsController_getContent",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "description": "ID du contenu",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contenu"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir un contenu spécifique",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/enroll": {
      "post": {
        "operationId": "LessonsController_enrollLesson",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inscription réussie"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "S'inscrire à une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/my/enrollments": {
      "get": {
        "operationId": "LessonsController_getMyEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des inscriptions"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir mes inscriptions",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/progress": {
      "post": {
        "operationId": "LessonsController_updateProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progression mise à jour"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour la progression",
        "tags": [
          "Lessons"
        ]
      }
    },
    "/api/lessons/{lessonId}/progress": {
      "get": {
        "operationId": "LessonsController_getLessonProgress",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "description": "ID de la leçon",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Progression"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir la progression dans une leçon",
        "tags": [
          "Lessons"
        ]
      }
    },
...
    "/api/admin/lessons/{id}": {
      "delete": {
        "operationId": "AdminController_deleteLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une leçon",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "EduLearn API",
    "description": "API pour la plateforme EduLearn",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InstructorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "avatar"
        ]
      },
      "LessonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "isFree": {
            "type": "boolean"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "duration": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "enrollmentCount": {
            "type": "number"
          },
          "instructor": {
            "$ref": "#/components/schemas/InstructorDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isEnrolled": {
            "type": "boolean"
          },
          "progress": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "thumbnail",
          "price",
          "isFree",
          "level",
          "status",
          "duration",
          "tags",
          "enrollmentCount",
          "instructor",
          "createdAt",
          "isEnrolled",
          "progress"
        ]
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à JavaScript"
          },
          "description": {
            "type": "string",
            "example": "Apprenez les bases de JavaScript"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://example.com/thumbnail.jpg"
          },
          "price": {
            "type": "number",
            "example": 29.99
          },
          "isFree": {
            "type": "boolean",
            "example": false
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner"
          },
          "duration": {
            "type": "number",
            "example": 120
          },
          "tags": {
            "example": [
              "javascript",
              "programming"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "price",
          "isFree",
          "level",
          "duration"
        ]
      },
      "UpdateLessonDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction à JavaScript"
          },
          "description": {
            "type": "string",
            "example": "Apprenez les bases de JavaScript"
          },
          "thumbnail": {
            "type": "string",
            "example": "https://example.com/thumbnail.jpg"
          },
          "price": {
            "type": "number",
            "example": 29.99
          },
          "isFree": {
            "type": "boolean",
            "example": false
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "beginner"
          },
          "duration": {
            "type": "number",
            "example": 120
          },
          "tags": {
            "example": [
              "javascript",
              "programming"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Module 1: Les bases"
          },
          "description": {
            "type": "string",
            "example": "Introduction aux concepts de base"
          },
          "order": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "title",
          "order"
        ]
      },
      "CreateContentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Introduction"
          },
          "type": {
            "type": "string",
            "enum": [
              "video",
              "text",
              "quiz",
              "document"
            ],
            "example": "video"
          },
          "content": {
            "type": "string",
            "example": "https://example.com/video.mp4"
          },
          "duration": {
            "type": "number",
            "example": 15
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "isFreePreview": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "title",
          "type",
          "content",
          "duration",
          "order",
          "isFreePreview"
        ]
      },
      "EnrollLessonDto": {
        "type": "object",
        "properties": {
          "lessonId": {
            "type": "number",
            "example": 1
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "stripe",
              "paypal",
              "mobile_money",
              "bank_transfer"
            ],
            "example": "stripe"
          }
        },
        "required": [
          "lessonId",
          "paymentMethod"
        ]
      },
      "UpdateProgressDto": {
        "type": "object",
        "properties": {
          "contentId": {
            "type": "number",
            "example": 1
          },
          "isCompleted": {
            "type": "boolean",
            "example": true
          },
          "timeSpent": {
            "type": "number",
            "example": 120
          }
        },
        "required": [
          "contentId",
          "isCompleted",
          "timeSpent"
        ]
      },
...
    }
  }
}